% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ay.R
\name{add_ay}
\alias{add_ay}
\title{Add academic year}
\usage{
add_ay(.data, term)
}
\arguments{
\item{.data}{your data}

\item{term}{a term field, where the fourth character is the season, and the second and third character represent year.}
}
\value{

}
\description{
Add academic year
}
\examples{
#df\%>\%
#add_ay(Term)
}
