% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_degree_level.R
\name{add_degree_level}
\alias{add_degree_level}
\title{Add Degree Level}
\usage{
add_degree_level(.data, degree)
}
\arguments{
\item{.data}{your data}

\item{degree}{field housing the degree code}
}
\value{

}
\description{
This simplifies degree codes to make all associate-level awards one category.
Adds a column named DegreeLevel
}
\examples{
#df\%>\%
#add_degree_level(Degree)
}
