% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ftpt.R
\name{add_ftpt}
\alias{add_ftpt}
\title{Add full-time/part-time}
\usage{
add_ftpt(.data, ftpt)
}
\arguments{
\item{.data}{your data}

\item{ftpt}{the column holding your various}
}
\value{

}
\description{
Add full-time/part-time
}
\examples{
#dataset\%>\%
#add_ftpt(Full_Part)
}
