% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_numeric_grade.R
\name{add_numeric_grade}
\alias{add_numeric_grade}
\title{add_numeric_grade}
\usage{
add_numeric_grade(.data, grade, credits)
}
\arguments{
\item{.data}{your data}

\item{grade}{the name of the column housing the letter grades you want to convert}

\item{credits}{the column housing the number of credits the course is worth}
}
\value{

}
\description{
add_numeric_grade
}
\examples{
#df\%>\%
#add_numeric_grade(Final_Grade,Credit_hours)

}
