% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_not_in.R
\name{filter_not_in}
\alias{filter_not_in}
\title{Title}
\usage{
filter_not_in(.data, OtherDf, columnInData, columnInOtherDf)
}
\arguments{
\item{.data}{your dataframe}

\item{OtherDf}{another data frame that you are comparing records with}

\item{columnInData}{the column in your original dataframe that will be the basis of the filter}

\item{columnInOtherDf}{the corresponding column found in the dataframe that you are comparing to}
}
\value{

}
\description{
When you are looking to filter out items that appear in another dataframe,
negating an \%in\% filter returns all records if you mistakenly refer to a column that
does not exist in the comparison dataframe.  This function protects against code
running smoothly and the risk of missing the mistake by throwing an error.
}
