% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_datatable.R
\name{output_datatable}
\alias{output_datatable}
\title{Title}
\usage{
output_datatable(df)
}
\arguments{
\item{df}{data that you are wanting to output a nice table for}
}
\value{

}
\description{
Title
}
\examples{
#output_datatable(df)
}
