% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentage.R
\name{percentage}
\alias{percentage}
\title{percentage}
\usage{
percentage(..., digits, symbol = NULL)
}
\arguments{
\item{digits}{the number of digits to round your percentage to}

\item{symbol}{the symbol you'd like to use, if any (generally \%)}
}
\value{

}
\description{
percentage
}
\examples{
#percentage(.74621,digits=2,symbol="\%")

}
