% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_better.R
\name{round_better}
\alias{round_better}
\title{Title}
\usage{
round_better(x, roundTo, dir = NULL)
}
\arguments{
\item{x}{number to be rounded}

\item{roundTo}{a number to round to "the closest of"}

\item{dir}{optional directional argument. "Up" always rounds up, "Down" always rounds down}
}
\value{

}
\description{
Title
}
\examples{
#round_better(7924,1000,"Up")
#round_better(7924,11)
}
