% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_nona.R
\name{sum_nona}
\alias{sum_nona}
\title{sum_nona()}
\usage{
sum_nona(...)
}
\arguments{
\item{...}{your data, generally just piped in}
}
\value{

}
\description{
take to sum, but ignore na values.  Just a means of reducing arguments.
}
\examples{
#df\%>\%
#sum_nona(Amount)
}
